Updating a the sp2 hbond score function -- "version 8" does not use the chi/BAH term for short-range backbone-backbone hbonds (unlike v7 which did). Default values for the amplitude and peak-to-trough height parameters adjusted.
Setting the minimum value for the cos( BAH angle ) to -1 and fixing a 
subtle bug associated with this change that had caused an unexpected 
integration test change:

0 + 0 + 0 < 0
for some values of 0. 

This caused some hydrogen bonds with energies ~0 to show up as hydrogen bonds
and trigger the (soon to be deprecated) bb/sc exclusion rule and throw out
perfectly good hbonds. Now, if one of the hbond parameters is outside of the
fade-function range AND is outside of the polynomial range, then the 
energy is set to MAX_HBOND_ENERGY + 1.0 instead of ~0.  

E.g.:  cosBAH = -0.5 w/ standard_params.  The fade function for cosBAH (fxH)
will evaluate to 0.0 and turn off the poly( AHdist ) and the poly( cosAHD )
terms.  It's own polynomial, poly( cosBAH ), has a lower-limit specified
of cosBAH >= 0.0 for it to be evaluated, so it will return 0.

In this case, the code will now return an energy of MAX_HBOND_ENERGY + 1
instead of fxH*poly( AHdist ) + fxH * poly( cosBAH ) + poly( cosBAH ) 
which should be zero but for some reason comes out to be slightly less
than zero.

No integration test changes.  All unit tests pass.



git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@44905 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: 2214222c6619346b29fc2e1a061726ca9dfa70ef