Adding code to Rosetta to aid in handling noncrystallographic symmetry.  This should be generally useful in cases where one wants model a system asymmetrically but wants local structural similarity between chains/segments maintained during minimization/packing.
This checkin adds a new constraint type, 'DihedralPairConstraint,' that restrains a pair of dihedrals to take the same value.  The checkin also adds a new function type, 'TopOutFunc,' meant to be used with these constraints, which behaves like a harmonic at small values, and becomes flat at the parameter 'limit_' (but is differentiable throughout).  To use these in scoring specify a non-zero weight for 'dihedral_constraint'.

This constraints can be specified on a per-torsion level using the ConstraintIO framework:
  DihedralPair   C 1 N 2 CA 2 C 2   C 101 N 102 CA 102 C 102  TOPOUT 0.01 0 10
The first four terms define the source dihedral; the second four the target, and the remainder the target function: for TOPOUTFUNC, the three parameters are <weight> <x_target> <limit>

However, this checkin also includes a parsable mover, SetupNCSMover, that greatly simplifies the task of setting up these constraints over many torsions.  For example, to constrain all the dihedrals of chain A (residues 1-100) and chain B (residues 101-200) to take the same value:
	SetupNCSMover setup_ncs;
	setup_ncs.add_group( "1A-100A", "1B-100B");
	setup_ncs->apply(pose);

Or using RosettaScripts:
<SetupNCS name=ncs bb=1 chi=1 limit=10>
	<NCSgroup source="1A-100A" target="1B-100B"\>
</SetupNCS>
Multiple sets of constraint may be specified by providing multiple 'NCSgroup' subtags.  In both cases if the chain ID is omitted then Rosetta numbering is used.




git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@45396 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: d8700dbf656b64aba6af77de48eede0ae1c12e38