Checking in code for a minimizable centroid scorefunction.
This checkin includes a smooth reparameterization of the following score functions:
	cen_env_smooth (equivalent to 'env')
	cen_pair_smooth (equivalent to centroid 'pair')
	cbeta_smooth (equivalent to 'cbeta')
	cenpack_smooth (equivalent to 'cenpack')

The scorefunctions were derived by fitting mixtures of gaussians to the current scorefunctions and using sigmoidal neighborcount smoothing for env and cenpack (as is done for the fullatom envsmooth term).  The pair term, which was originally computed in wide bins, was recalulated at finer sampling over a set of 10k PDBs.  Because of the reparameterization, the scores are different than their non-differentiable counterparts, so they are implemented as new score terms.  The agreement to the original nonsmooth variants on a set of ~500 native structures of similar size is quite high for env and cbeta (correlations of 0.99 for env and 0.97 for cbeta) but lower for cen_pair and cenpack (~0.5).

Additionally this code also includes a smoother reparameterization (smoothed in both donor/receptor angle and distance) of the centroid backbone hbond term, 'cen_hb', meant for use in minimization of centroid structures.  On the same set of 500 structures, the correlation with (hbond_lr_bb+hbond_sr_bb) is 0.98.

The reparameterized scorefunction is checked into the database as 'score4_smooth.wts'  A variant, 'score4_smooth_cart.wts' also turns on the harmonic bond-angle / bond-length term for Cartesian minimization (or flexible bond-length & bond-angle minimization).  The parameter files for these new scorefunctions are checked into the database in scoring/score_functions/centroid_smooth/.

Finally, derivatives for 'rg' have been added (the score remains unchanged); this causes integration test changes for the following 8 tests, which minimize against the centroid scorefunction (via looprelax which performs a centroid linmin):
 density_refine
 density_refine_symm
 flexpepdock_abinitio
 full_length_model
 kinematic_looprelax
 ligand_dock_7cpa
 loop_modeling
 mr_protocols

Also the following tests that fail whenever a scoreterm is added will fail: database_jd2_io, features





git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@45618 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: f09585ba657a6d5c707ab4f86779716c4c727338