changing list of string_value pairs in Job to map.
IMHO it makes not much sense to have multiple entries with same name. 
Rather it is actually useful to be able to overwrite entries during the run of a protocol. 
I looked at common usages of this feature and haven't seen any that make a functional use of the possibility to store
multiple values with same name. Now it is changed. This causes a lot of cosmetic changes to integration tests, 
since the string-values are now sorted by name. 





git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@46119 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: 6ceb583c23b08f7b6050059c99790384d24e000c