1) adding support for simulated tempering to Metropolis hastings framework. 2) splitting TrajectoryRecorder into Baseclass and two derived classes. now the SilentTrajectoryRecorder ( should be renamed soon to JobOutputterRecorder ) uses jd2:output_intermediate_pose() for output and the PDBTrajectoryRecorder writes poses to a multimodel PDB file as before.
important difference: we don't write the initial pose to the trajectory. This has the simple reason that the initial pose might have
very different set of score-entries and thus would mess up columns etc.

this change is seen in the integration test: MODEL 0 is missing. but the remaining models are the same.



git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@46122 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: 5cd1045465cb191830d8b3cef77b980fbeaf92e8