Introducing a runtime_assert to catch cases when symmetric repack happens without the use of a SymmetricRotamerSet_. This can happen if the symmetry:symmetry_definition flag is not set during a symmetry run.
I have run the unit and integration tests. On my machine the ligand_database_io integration test reported a difference for the output pdb in the residue score output. This was in the 9th decimal. This may not be replicated on the test server but if it does it should be of no significance. For non-symmetric runs (like ligand_database_io) the code adds one additional line, a check if a pose is symmetric, which should have no impact on the run (this test function is excecuted hundreds of times in the code elsewhere).



git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@46735 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: 655752fb5a6368e93346ef193e8004be5b85bf56