Removing get_loop_file_name and get_loops_from_file as free-floating functions in Loops.cc.  Instead Loops now has two additional constructors: one takes a boolean (true will read in the standard loops_file option) and another that allows a filename to be specified (because everyone seems to use the loops code to read in all sorts of stuff).  To handle this, the Loops constructors all call a designated initializer that is responsible for setting up the rest of the class.
The motivation for this is that the loops class is now responsible for the loop_file data.  Instead of telling the 
instance to load data from disk, you indicate at construction if you want that instance to do that or if you plan to set it up
through code.

The loop_file name is stored as a data member in the class, and the mutator for it triggers a reload of the loop_file
from disk.  All unit tests pass and no integration tests change.



git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@46993 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: 6374513dae4110f8f9085769f5625f3f6e31813c