Changing the definition of the loop in the LoopAnchorFeaturesReporter.  The anchors are now considered as a part of the loop when calculating the loop length.  This causes a change to the features integration test.
This commit also introduces a new tag option for the LoopAnchorFeaturesReporter: use_relevant_residues_as_loop_length. When used in conjunction with a TaskOperation that selects a single loop, this is useful for 
reporting features of individual loops when the start and stops are well-defined.



git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@47453 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: c86a67895e0a2ca4ddcae7fc12c3611ac7f1e597