This commit consists solely of two changes to the build system.
1) The big change consists of removing some wonky logic
in rosetta/rosetta_source/tools/build/setup_platforms.py in the select_compiler_version function.  Originally, this
function would take a requested compiler version and return an actual compiler version.  By default, the requested
version is "*".  The function would then determine the compiler version and set actual to the version of the compiler
that scons could detect.  Before this value was returned, it was reset to "*" for unknown reasons.

This commit removes the resetting step.

2) In OS X, gcc/g++ are actually using llvm-gcc/llvm-g++.  This commit makes scons explicitly call llvm-g++-4.2 when 
g++-4.2 is detected by the system.  This is necessary on 10.7 becuase g++-4.2 doesn't exist.  Additional changes have 
been made to the build rules for clang on OS X that will result in the removal of linker warnings.

I have tested these changes on a mac with both gcc and clang. No test changes expected (unless of course new paths to
the libraries cause EVERY integration test to fail).

SCons is really hard to work with.  Many thanks to Sam DeLuca and Sergey Lyskov for all of their help with tracking this
stuff down.



git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@47578 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: d101f5186864448e7d72b5f9dcddf65200ddf706