Removing protocols.src.settings.
protocols.src.settings is no longer read by scons and modification to this file will have no effect on the build. Please be careful when merging not to add it back as this can lead to confusion for developers.


git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@47775 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: 18608dcf21da027a99507c60d34e3956e165fdeb