the C_terminal part seems working. It actually matched R2 antibody result
but the N_terminal part is weird. The problem is from a bad grafting.
After superimposing the template pose stems on the top of framework stems,
the residues were directly copied. 
Assuming H3 is [a,b]
the  a-1 residue is from the loop template, not from the framework.
I am not sure this is a problem or not, because CCD should eventually take
care of it. But need to check this later for sure!



git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@48001 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: a7de066ab8ac5265e583fc8fdad822d0092f565d