Add command line patch loading
Although people have had the ability to add new residue types which aren't in the database with the command line options with -extra_res_fa and -extra_res_cen, so far they haven't had the ability to add patch files which aren't in the database. (e.g. if there's a patch for the residue specified with -extra_res_fa).

This commit adds two options, -extra_patch_fa and -extra_patch_cen, which add user-specified patch files to the list of patches to load in addition to the ones in the database.

No test changes observed, except for the new unit tests for ResidueTypeSet.

git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@48386 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: 527d160640f8932641d3bf6e5b3a1ccd0798c545