A bunch of structural changes to qsar/scoring_grid.  GridBase is now a simple interface class, and SingleGrid is what GridBase used to be.  This was done so that a new type of grid, a "metagrid" could be created.  A MetaGrid is a scoring grid composed of other scoring grids.  A new grid, "ChargeGrid" was created, based on the electrostatic grid implementation described in Goodfords GRID paper from 1985.  ElectroStaticMetaGrid is a set of ChargeGrids spanning a range of charges.
I also added a function to numeric/util, give it a vector1 of numbers, and a query, and it will return the value in the vector1 which is closes to the query. It does this by sorting the vector and then doing a binary search. There is a new unit test to test this function.

The relax_allatom_w_cst integration test failed with that weird IG change we've been seeing, but this code doesn't touch anything that I could imagine causing such a change so i dont think it's real

git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@48432 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: 4745202f3047ce6a05b7a510666bf355dec55528