Adding new RosettaScripts functionality to devel/matdes. No changes expected.
All unit tests passed; two integration tests are different: one was reported previously and is just cosmetic (tracer output: "Creating tricubic splines for the backbone-dependent non-rotameric chi scores") and the other was actually a problem with my clean reference build as well in which one of the tests, darc_score, did not run. As I said, this was a problem with the fresh checkout as well and thus should not have to do with any changes that I have made.



git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@48671 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: 6afe392ab7bea16af55cd918b9ad4bb94d4f4390