Add option to change output scorefunction in RosettaScripts.
RosettaScripts now can take a top-level <OUTPUT /> tag,
that can take a scorefxn parameter which specifies which
scorefunction to use in the final scoring before output.

For example:

<ROSETTA_SCRIPTS>
  <SCOREFXNS>
     <myscore weights=score12prime>
	<Reweight scoretype=fa_pair weight=0 />
	<Reweight scoretype=hack_elec weight=0.5 />
     </myscore>
  </SCOREFXNS>
...
  <OUTPUT scorefxn=myscore />
</ROSETTA_SCRIPTS>

will cause the output PDBs to contain the final scoring information of the
"myscore" scorefunction, rather than the commandline scorefunction. If the
OUTPUT tag is not specified (or if the scorefxn parameter is missing) then
the final scoring of the pose will be the default commandline scorefunction.

Additionally, a new predefined scorefunction named "commandline" has been
added to RosettaScripts, correspoding to whatever default scorefunction which
was specified with the commandline options (subject to caveats discussed below).
(Note that while it's now availible by name, it's use in scripts is not really encouraged -
you're probably better off explicitly specifying scorefunctions used in the XML.)

Cosmetic tracer changes to RosettaScripts integration tests (additional
scorefunction loading and/or score function loading at a different time):

database_jd2_io
features
filter_termini_near_interface
inverse_rotamer_remodel
kinemage_grid_output
ligand_database_io
ligand_dock_grid
ligand_dock_script
ligand_motif_design
metropolis_hastings
posttranslationalmod_io
ppk
rosetta_scripts_setup
rotamer_probability
rs_flexbbmoves
rs_loophash
score12_docking
score_only_silence
startfrom_file
Enzrevert_xml
rosetta_scripts_loops

Cosmetic tracer changes as well as output changes to the following:

hotspot_graft
place_simultaneously

In the new output an extra backbone_stub_constraint and/or a
res_type_constraint term is added to the scorefunction because the associated
XML scripts use the SetupHotspotConstraints and/or SequenceProfileMover movers
(via the APPLY_TO_POSE tag), which in their parse_my_tag functions
indiscriminately go through and add backbone_stub_constraint/
res_type_constraint to all of the defined scorefunctions in the datamap.

This means that the default output scorefunction (which is now stored in the
datamap under the "commandline" name) also get those scores turned on.
I'm putting this in the "it's not a bug - it's a feature" category, as due
to the construction of the two movers, it's impossible to use a scorefunction
without those terms in an XML with them. The previous way masked the fact
that *all* scorefunctions used within the XML had this applied.

If the new behavior is not desired, I think the best way of dealing with it
would be to change the SetupHotspotConstraints and SequenceProfileMover
to be more specific in which scorefunctions they modify the weights for.

No unit test changes observed

git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@49383 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: 776777c4641c796bc5f370f57a40f361eafa5baa