Analytic etable evaluation now available, controlled by the flag -analytic_etable_evaluation. The analytic form was first derived exactly from the tabular form, but was refined in a very minor way to remove a derivative discontinuity at the start of the fasol smooth-to-zero spline. When activated, it reduces the etable's memory footprint by 108 MB (with the standard fa_max_dis of 6.0 A) and by 243 MB if fa_max_dis is extended to 9.0A.  Because the function is computed analytically instead of using table lookups, the derivatives are exact.  This improves minimization, especially in cases of collisions where previosly 'Inaccurate G!' messages would have been produced, but it also improves regular minimization so that fast_relax produces slightly lower energies.
The analytic evaluation is ~30% slower than the tabular evaluation.  For this reason,
a 16.5 second relax trajectory for ubiquitin takes 17.5 seconds using the analytic
evaluation.

The lookup-table evaluation scheme remains the default. Analytic evaluation can be
turned on from the command line, in which case the large lookup tables are not 
allocated (saving memory) but meaning that the tabluar form cannot be used.
Alternatively, the analytic lookup scheme can be activated programatically using the
EnergyMethodOptions object, allowing a protocol writer to activate the analytic
form for portions of their protocol (e.g. minimization) while falling back on
the table-lookup form in other, less sensitive phases of the protocol (e.g. packing).

To incorporate this change, the surrounding code in the BaseEtableEnergy object
had to change signficantly and this has resulted in small numeric changes that
produce integration test changes.  The code has been extensively tested. 
The new table-lookup code matches the values produced by the old table-lookup
code to high precision, so the resulting integration test changes are not worrying.

This commit also alters the spline that ramps the Lennard-Jones atractive
component of the energy function to zero in the range 4.5 -- 6.0: for 
Br/Br and Br/I interactions, the sum of the van der Waals radii is
slightly greater than 4.5 A, so that the spline starts ramping toward
zero before the minimum-energy distance.  I'm now starting the spline
at the maximum of the van der Waals sum or 4.5 A.  I'm bumping the version
number for the EtableEnergy, which will produce a score function test change.
 





git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@49443 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: 045979cb7ffa023a917203dc6a004ac75a246e60