Check in "sasapack" and "avge" scores for evaluating decoy/design quality. Two new files (protocols/sasa_scores/sasapack.hh;.cc); the only changes to existing code are addition of two new extra arguments (with default values) to calc_per_atom_sasa in core/scoring/sasa.cc. Compiles for me; no test changes observed. Additional checkins will include apps for evaluating/refitting the scores and python code for getting the distributions, creating new database files. Need to update the rosetta database to get the new input files (in scoring/sasa_scores/).
sasapack is a simple-minded packing score that compares residue
solvent accesible surface area with a small probe (0.5 A) and with a
water-sized probe (1.4A). Based on distributions of these two values
for residues in a large dataset of native pdbs, poorly packed residues
are identified as those for which the SASA-0.5 is larger than you'd
expect based on the SASA-1.4 and the residue type. Updates relative to
the old r++ version: use rosetta's LJ radii for the sasa computation;
better fitting scheme for getting the SASA distributions from the
natives, including an app and python code to easily re-fit the
distributions (these will be checked in shortly). Also now computing
just the normalized SASA-1.4, which for each residue is the difference
between the SASA-1.4 value and the average SASA-1.4 value for that
residue type. This helps to diagnose one way that decoys can cheat and
get a better sasapack score, namely increasing the SASA-1.4
values. Plotting the normalized SASA vs the sasapack looks useful for
analyzing decoy sets. There are probably better ways of getting a true
normalized SASA score, though: protein length, topology, and ss
content could be used. Note that there is no reason to expect that sasapack
would be at all superior to Will's packing score, in fact, quite the
reverse. I'd probably expect them to be pretty well correlated. 

The "avge" (average-E --> AvgE --> avge) score looks at a residue's
total energy relative to energies for that residue in native
structures, normalized by the SASA-1.4 value. This is like Brian's average
energies that Rosetta used to write to PDB files, which were
normalized by number of neighbors. Important gotches here are that the
exact scorefunction used to evaluate the natives should be used for
scoring the decoys, including any flags that would modify the
values. Right now "score12prime" was used with no cmdline flags for
scoring natives; in evaluating a pose, the pose will be rescored with
a newly created "score12prime" scorefxn, but the energies will only be
comparable if no score-modifying cmdline flags are present. As with
sasapack, I will check in an app and python code for re-fitting the
distributions using different pdb sets, scoring functions, -correct,
etc. For now, the natives were not refined at all, so fa_rep, fa_dun,
pro_close, omega, p_aa_pp for gly are not included in the
energies. Backbone hydrogen bonds are also not included in the residue
total energies. So for now the score is mainly focusing on LJ, LK,
sidechain hbonds.





git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@49994 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: 4ed59d0c05322a15aaef1fe743ca837af335a6c9