continue refactoring the AntibdoyInfo class.
The fold_tree utilities in the AntibodyInfo was able to change the pose. Finally changed all the functions like this, so that it can only return values, and the function itself can be constant.

As always, the change of AntibodyInfo class require the changes in other antibody related classes.

the log files of the 3 integration tests are expected to have some cosmetic changes  




git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@50932 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: f60f982d8913df07af7130110cadcc65bc52fc81