Fix a bug in r49036 so uuid data is written to sqlite3 databases with type BLOB rather than TEXT.
With SQLite, apparently, it is possible to put any type in any column:
http://sqlite.1065341.n5.nabble.com/Loss-of-Binary-Data-in-Dump-File-td59757.html

The problem with putting the uuid's in as text is that they are dumped
as text literals (which I can't even show you because SVN thinks its
badly formated UTF-8 data) rather than as a clean blob literals like this:

   X'61C7D84133974A2F90EEDB5B8BEEE9E2'

Then if the dump is read back into the database, the
presense of a null character will truncate the value.  This can lead to
data inconsistency (to say the least), for example when reporting
features using MPI and then using
rosetta_tests/features/sample_sources/merge.sh to merge the database
parts together.

This causes changes in all integration tests that dump feature
databases for textual comparison. No unit test failures are observed.


git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@53150 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: d24ff7cf58aad878daeafe5a447244284edbcb71