Added a blueprint option to LayerDesign, which allows specification of secondary structure through a blueprint file rather than DSSP (default=use DSSP)
Added options to LayerDesign to enable specification of which residues are referred to in layers. For example, a layer could be defined as a set of residues marked as "designable", "repackable", or "fixed" by a layer subtask.

Added options to LayerDesign to enable specification of an action to be taken for each layer. For example, a layer could be set to "design" (use layerdesign to design), "omit" (layerdesign doesn't touch the residue), or "nodesign" (layerdesign explicitly sets the residue to repack only or fixed rotamer)

No integration or unit tests expected.

git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta/@54633 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: 4ccd9427b40725e9b4a129c78be59badd9d77984