Clean the antibody_graft integration tests
1. Make the optios of -antibody::graft_l1, -antibody::graft_l2,
-antibody::graft_l3, -antibody::graft_h1, -antibody::graft_h2,
-antibody::graft_h3 be default
2. Turn on the benchmark option, so that the stem optimization can take
much less cycles and the integration test can be really fast
3. Change the example in the integration test, the old one used the
output from the perl script. So I had to turn on the "R2_Perl" flag in
the past. Now the default is "R3_Python"

Did not run the unit test, as the only changes are some default option
changes in the "GraftCDRLoopsProtocol.cc" and "GraftedStemOptimizer.cc".
One integration changes happened, which is the antibody_graft
integration test that I worked on
The reason for the changes are:
1. the input a few PDBs have been changed
2. the options in the flag file has been changed
3. results are changed due to 1 and 2.

Old repository SHA1: 4a93479661deeffbb55784e6f4f08ff25214e79f