Added the last flag -antibody:h3_no_stem_graft so that the H3 stem can still be the stem from the Framework template. But be sure to have idealization of the 0th, and n+1st (assuming H3 is 1-n) residue bofore H3 loop modeling, so that the bond length can be natural.
Again, no C++ source code change, the integration tests changed due to
the changes in the flag files

----
An additional 7 commit messages were squashed into this commit:
--
remove a bunch of useless flags in the flag file

Again, no C++ source code change, just the changes of the flag file
--
R3_Python is default, remove this flag to reduce confusion

added a flag to include the native structure

No C++ source code was changed in this commit. All the changes are due
to the flag changes
--
This is another commit to change the input files of the antibody_graft

The old example used R2_Perl outputs as inputs for grafting, now we use
R3_Python results. Because the old H3.pdb for R2_Perl will not work, I
had to make a new antibody target. Therefore I had to update L1.pdb,
L2.pdb, L3.pdb, H1.pdb, H2.pdb, H3.pdb. and I need to change the flags
to R3_Python.

This integration tests will change because of all the reasons above.
But, not a signle C++ source code was changed in this commit, therefore,
it is guranteed that the changes are due to the input files in the
integration test itself.
--
remove the unnecessary directory input and the files in there. This
integration test should change again only because of these files changes
--
Turn on the -run:benchmark flag in the flags file, this integration
tests got changed. Everything changed, the flags file, the output decoy,
the intermedidate files, the log files, because it does less cycles of
stem optimization now
--
added the option of turning on benchmakr for the antibody_graft, so that
it will run much less cycles for the speed issue in integration tests.

This flag has not been turned on yet, therefore, no integration tests
changed.
--
make the options of graft_l1, graft_l2, graft_l3, graft_h1, graft_h2,
graft_h3 all default to be true in the source doe. These options can be
deleted in the flag file of the antibody_graft integration tests. The
antibody_graft integration tests changed only because of the flag file
changed
--

Old repository SHA1: 86a0b17ad3e8bf35bc3cde6b0c1ed373d31fab19