Repackwithoutligand Fix minor bug with RMSD input computation.
The files have been modified such that the RepackWithoutLIgand filter bails out if no atoms are specified for the interface - now it returns a NaN and the filter does not fail. Furthermore, a minor bug has been fixed where the length of the pose with ligand and without ligand now have the same length. This means that the application can be run within debug mode.

Old repository SHA1: 152cc6be409e49be28b0b5856c0fc14909940068