Revert 289429e724 / r55348 for unexpected integration test changes.
----
An additional 1 commit messages were squashed into this commit:
--
Fix bug in DockMinMover.

MonteCarlo objects need to be initialized appropriately, otherwise if the first move is a failure, it will reset the pose to an incorrect state.

This fixes a bug that was noticed on the RosettaCommons Forum.
--

Old repository SHA1: 21d86e1b71148cacb2ef1131a3d61ecb716a2b64