Adding new matcher option to reduce memory usage.
Adding a new flag, -build_round1_hits_twice, that uses a new strategy
for generating hits.  In round 1, hits are generated at one build
point at a time, and placed directly into the occupied-space hash.
They are then discarded before going on to the next build point.
Round two runs as usual.  Then, after round 2 completes, the hits
from round 1 are regenerated, but only the hits that land in the 6D
bins in which hits from round 2 are stored.  This is often a very
small minority of all the hits that were generated in round 1.

To use this, neither of the first two geometric constraints (the ones
used in rounds 1 and 2) can rely on secondary matching.

----
An additional 2 commit messages were squashed into this commit:
--
fixing tracer output message stating how many round-1 hits have been regenerated
--
First attempt at a new matcher strategy that should reduce memory use

The idea is that, if you want to do really heavy sampling, then round 1
builds waaaay too many hits, the vast majority of which get thrown out at
the end of round 2.  So why not only keep the location of the round 1 hits
and not the hits themselves?  The round1 hits are marked in the occupied
space grid (and then discarded). This info is used to prune the hits that
would be generated in round 2 as normal.  The round 2 hits then can be
used to trim back the regions in the occupied space hash that would lead
to viable matches.  Then the round 1 hits can be regenerated.

There's no reason why more constraints couldn't be processed in this manner
but for now, I'm just implementing it for round 1.
--

Old repository SHA1: 0ecd1880458aae216ab63536d4ee76d4e5f51ec0