New functionality in the matcher to let you set constraints for separate atoms on symmetric residues
The new DESYMMETERIZE command that can be included as part of the
matcher's ALGORTIHMO_INFO block in a matcher constraint file
will tell the Dunbrack rotamer library to treat the symmetric
residues (D/E/F/Y) as if they are asymmetric.  I.e., it will
double the sampling by adding 180 to the terminal chi angle
for these amino acids.  This now lets you set constraints
so that OD1 can be connected to one partner, and OD2 to anther.

E.g., the following pair of matcher constraints would say
that OD1 should contact a zinc residue, and that OD2 should
contact either an ASN or GLN residue.

CST::BEGIN
  TEMPLATE::   ATOM_MAP: 1 atom_name: ZN V3 V1
  TEMPLATE::   ATOM_MAP: 1 residue3:  ZNX

  TEMPLATE::   ATOM_MAP: 2 atom_name: OD1 CG CB
  TEMPLATE::   ATOM_MAP: 2 residue3:  ASP

  CONSTRAINT:: distanceAB:   2.10   0.15  40.0  0    1
  CONSTRAINT::    angle_A:   109.5  20.0  40.0  360. 1
  CONSTRAINT::    angle_B:   120.0  30.0  40.0  360. 1
  CONSTRAINT::  torsion_A:  -120.0  20.0  40.0  360.
  CONSTRAINT:: torsion_AB:     0.0  20.0  40.0   20.
  CONSTRAINT::  torsion_B:     0.0  10.0  40.0  180. 1

  ALGORITHM_INFO:: match
    CHI_STRATEGY:: CHI 1 EX_FOUR_HALF_STEP_STDDEVS
    CHI_STRATEGY:: CHI 2 EX_ONE_STDDEV
    DESYMMETERIZE
  ALGORITHM_INFO::END

CST::END

CST::BEGIN
  TEMPLATE::   ATOM_MAP: 1 atom_name: OD2 CG CB
  TEMPLATE::   ATOM_MAP: 1 residue1: D

  TEMPLATE::   ATOM_MAP: 2 atom_type: NH2O,
  TEMPLATE::   ATOM_MAP: 2 residue1: NQ

  CONSTRAINT:: distanceAB:    2.9   0.30    20.00   0    1
  CONSTRAINT::    angle_A:  120.00  20.00   0.00  360.00 2
  CONSTRAINT::    angle_B:  120.00  30.00   0.00  360.00 2
  CONSTRAINT::  torsion_A:    0.00  30.00   0.00  180.00 4
  CONSTRAINT::  torsion_B:    0.00  30.00  25.00  180.00 1

  ALGORITHM_INFO:: match
    SECONDARY_MATCH: UPSTREAM_CST 1
  ALGORITHM_INFO::END
CST::END

----
An additional 3 commit messages were squashed into this commit:
--
Connecting the desymmeterize functionality with the Matcher initialization.
--
Unit test for the desymmeterization code in the matcher.
--
Attempting to add functionality to desymmeterize the symmetric amino acids in matching.

That is, I want to be able to set a constraint where OD1 on ASP contacts one residue,
and a second constraint where OD2 contacts a different residue.  Currently, the only
way to do this is to define two constraints, each one of which makes a contact to
an OOC atom.  This leads to the possibility that both contacts are made to the
same oxygen atom.
--

Old repository SHA1: 9f2044aed1cbec6f12ce2a0d02d973dc5424c3fd