Fix issues with Hydroxyproline loading.
Because the hydroxyproline patches use "HD1" as the atom name for the hydrogen on the hydroxyl group,
some input files from non-Rosetta programs can misclassify prolines as hydroxyprolines, as "HD1" is an alternative
 atom name for the hydrogen attached to the delta carbon. When loading the PDB, Rosetta will see the HD1 name and
conclude that hydroxyproline is a "better" match than regular proline.

This commit changes the name of that hydrogen to the alternate name "HOD" to avoid this confusion.

Old repository SHA1: 18ef07b505e522d9746831ff28e0472265b5a974