Fixing bug #290
Fixing two problems in the analytic version of the etable, one serious, the other
considerably more minor.

1) Fixing the definition of the close-range (shorter range) spline that ramps the
solvation term to a constant value below the van der Waals distance for heavy atoms
that's used by the analytic etable evaluation machinery.  Previously, the distance
given by the bin "E2" was given instead of the distance for bin "E1."  The rub is
that while the function was continuous, the derivative was discontinuous at E1.

2) Fixing the distance at which the lennard-jones minimum is taken.  Previously,
to conform to the etable, the distance was taken as the distance that has the
table entry (bin boundary) with the smallest value.  That is, the minimum value
is given by the sum of the van der Waals radii.  But this sum may not have fallen
on a bin boundary before, and so the minimum was taken as whatever bin produced
the minimum value.  There's really no reason to preserve this inaccuracy, so,
while I'm fixing bug #290, I'm going to change this behavior as well.

Unit tests are updated to reflect this change.  Almost all integration tests
will change since the etable is now going to behave differently.

Thanks, Matt, for finding this bug!

----
An additional 1 commit messages were squashed into this commit:
--
Three more utracer updates that go along with the previous commit
--

Old repository SHA1: 234478ca15aeb538c9e83f442676f7985cc770a4