Fixing an issue with the bumpcheck in kinematic closure
This commit removes the (recently added) backbone H bumpcheck -- if hbonding, these atoms will be closer to their donor/acceptor than the bumpfix allows.
This can lead to loops that are impossible to close (e.g. 2rn2 from the benchmark set): if the backbone NH of the first loop residue hbonds with an atom in the fixed part of the protein, these will always be too close to be considered acceptable (the NH position of the first residue is not changed during KIC, sampling starts at CA).

No local integration test changes observed, however the KIC-related tests might fail on the server.

----
An additional 6 commit messages were squashed into this commit:
--
merging in master

the plan is to comment out the bumpcheck for NH -- if they're involved in hbonds they will be closer than the bumpcheck allows, but that's a feature, not a bug
--
commenting out debug statement again, I'm not sure they are still accurate, would first need to trace exactly which atom we're dealing with in each step
--
some extra debugging output
--
reverting comment out -- 2rn2 does work when the H isn't checked, so that most likely is the root of the problem
--
next attempt -- now commenting out the NH check for alpha amino acids...
--
commenting out the bb NH check as we're concerned this is why 2rn2 stopped working
--

Old repository SHA1: fd29de8956ecd00a267a89a151e54f6cc3bb735e