ReportFilter can now checkpoint its data. See wiki for details
----
An additional 4 commit messages were squashed into this commit:
--
DesignInterfaces bugfix
--
fixing broken build
--
Improving the logic for ReportFilter. This can anow be called within a protocol at the step in which you want to save the information for any other filter. The value ReportFilter will report at the end of the run is the value of the filter at the step in which you called it. useful for benchmarking improvements in protocols
--
DesignInterfaces operation had an unimplemented function definition
--

Old repository SHA1: 5d605d8493c2a5fe2fc5c451fd4a1eaa34085068