Adding a new description for a ResidueType: its interchangeability group.
The packer will look at all the residues types in the a given interchangeability
group to decide which ones to allow.  This functionality replaces the previous
"name3" string which is overloaded for PDB I/O and is unnecessarily restirctive
in length.

The interchangeability group can be specified in the residue topology file,
but if it is not given, then it will be set from the IO_STRING field (the name3).

New PatchOperation classes for working with interchangeability group data.
Moving PatchOperation class definitions from the .hh file to the .cc file.

The "NC" ResfileOperation now uses the interchangeability group to append
non-canonical amino acids (and other not-canonical-amino-acid residue types)
to the ResidueLevelTask.

Updating the interchangeability groups for the residue types defined in the
fullatom/residue_types/rosetta_specific directory (which previously all
shared the name3 of "XXX").

Updating the patch files that used the SET_IO_STRING command so that they
now also update the interchangeability group.

----
An additional 5 commit messages were squashed into this commit:
--
Adding the SET_INTERCHANGEABILITY_GROUP command to all patch files that
use the SET_IO_STRING command; this effectively restores previous behavior.

Thanks, Doug, for making sure this happened!
--
Moving PatchOperation classes to the .cc file.  Adding
two new PatchOperations; SetInterchangeabilityGroup_String
and AppendInterchangeabilityGroup_String for working
with the interchangeability_group data in ResidueType.
--
Setting interchangeability group for the rosetta_specific
residue types in fa_standard to something meaningful
(XXX is not a meaningful group).
--
Adding interchangeability_group_map_ data to ResidueTypeSet.
Now PackerTask::add_noncanonical_aa reads from the
interchangeability_group_map_ instead of iterating across
all ResidueTypes in the ResidueTypeSet.
--
Adding data to class ResidueType and functionality to residue_io.cc
to describe an interchangeability group for each ResidueType so that
it can say it is functionally interchangeable with any other
ResidueType that's in the same group.  This will help avoid the
problem that name3 is too limiting in specifying which ResidueTypes
should be allowed in the packer.
--

Old repository SHA1: d16b60c9d6d788b88a7f91cf9cc7f5940c9a6662